move static function that retrieves all follows and pending follow requests to the table that displays it
